+++ c-ext/constants.c
@@ -89,4 +89,5 @@ void constants_module_init(PyObject* mod) {
 	PyModule_AddIntConstant(mod, "STRATEGY_LAZY2", ZSTD_lazy2);
 	PyModule_AddIntConstant(mod, "STRATEGY_BTLAZY2", ZSTD_btlazy2);
 	PyModule_AddIntConstant(mod, "STRATEGY_BTOPT", ZSTD_btopt);
+	PyModule_AddIntConstant(mod, "STRATEGY_BTULTRA", ZSTD_btultra);
 }