+++ tests/test_data_structures_fuzzing.py
@@ -34,7 +34,8 @@ s_strategy = strategies.sampled_from((zstd.STRATEGY_FAST,
                                         zstd.STRATEGY_LAZY,
                                         zstd.STRATEGY_LAZY2,
                                         zstd.STRATEGY_BTLAZY2,
-                                        zstd.STRATEGY_BTOPT))
+                                        zstd.STRATEGY_BTOPT,
+                                        zstd.STRATEGY_BTULTRA))
 
 
 @make_cffi